Session data now part of Client/Server in preperation for more complex game sessions. Project-wide code cleanup: no more compile warnings!

+++ src/net/fe/network/message/ClientInit.java
@@ -8,8 +8,8 @@ import net.fe.network.Message;
 public class ClientInit extends Message {
 	private static final long serialVersionUID = 2559169995718261494L;
 	public byte clientID;
-	public HashMap<Integer, Player> players;		// Players already in the lobby
-	public ClientInit(int origin, byte clientID, HashMap<Integer, Player> players) {
+	public HashMap<Byte, Player> players;		// Players already in the lobby
+	public ClientInit(byte origin, byte clientID, HashMap<Byte, Player> players) {
 		super(origin);
 		this.clientID = clientID;
 		this.players = players;
