+++ src/net/fe/network/message/EndGame.java
@@ -10,7 +10,7 @@ public class EndGame extends Message {
 	 */
 	private static final long serialVersionUID = 654326008504474145L;
 	
-	public EndGame(int origin, int winner) {
+	public EndGame(byte origin, int winner) {
 		super(origin);
 		this.winner = winner;
 	}
